# backend/src/docs/swagger.yaml
basePath: /
definitions:
  handlers.AddDepositRequest:
    properties:
      amount:
        example: 10000
        type: integer
      reference:
        example: BANK-TX-12345
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - amount
    - user_id
    type: object
  handlers.AddDepositResponse:
    properties:
      ok:
        example: true
        type: boolean
      transaction_id:
        example: TXN-1234567890
        type: string
    type: object
  handlers.AddLoanRequest:
    properties:
      amount:
        example: 100000
        type: integer
      borrower_id:
        example: 1
        type: integer
      duration:
        example: 12
        type: integer
      interest_rate:
        example: 12.5
        type: number
      reason:
        example: Home renovation
        type: string
      status:
        example: Approved
        type: string
    required:
    - amount
    - borrower_id
    - duration
    - interest_rate
    type: object
...
  handlers.AuditTransactionsResponse:
    properties:
      blockchain_verified_count:
        example: 3
        type: integer
      blockchain_verified_percentage:
        example: 100
        type: number
      total_amount:
        example: 65000
        type: integer
      total_count:
        example: 3
        type: integer
      transactions:
        items:
          $ref: '#/definitions/handlers.AuditTransactionItem'
        type: array
    type: object
  handlers.BlockchainStatusResponse:
    properties:
      blockchain_health:
        example: healthy
        type: string
      blockchain_integrity:
        example: true
        type: boolean
      last_block_number:
        example: 15000
        type: integer
      last_verification_timestamp:
        example: "2025-12-05T11:45:00Z"
        type: string
      pending_verifications:
        example: 0
        type: integer
      total_transactions:
        example: 1500
        type: integer
      unverified_transactions:
        example: 0
        type: integer
      verification_rate:
        example: 100
        type: number
      verified_transactions:
        example: 1500
        type: integer
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  handlers.FinancialSummaryResponse:
    properties:
      blockchain_integrity:
        example: true
        type: boolean
      last_blockchain_verification:
        example: "2025-12-05T10:30:00Z"
        type: string
      total_assets:
        example: 1500000
        type: integer
      total_deposits:
        example: 1000000
        type: integer
      total_interest_earned:
        example: 75000
        type: integer
      total_loans_disbursed:
        example: 500000
        type: integer
      total_loans_outstanding:
        example: 300000
        type: integer
      total_loans_repaid:
        example: 200000
        type: integer
      total_members:
        example: 150
        type: integer
      total_profit:
        example: 50000
        type: integer
    type: object
  handlers.HomeResponse:
    properties:
      blockchain_integrity:
        example: true
        type: boolean
      dividend_expected:
        example: 5000
        type: integer
      role:
        example: member
        type: string
      total_assets:
        example: 1000000
        type: integer
      total_loans:
        example: 500000
        type: integer
      total_profit:
        example: 50000
        type: integer
    type: object
...
  handlers.InterestRateResponse:
    properties:
      rates:
        items:
          $ref: '#/definitions/handlers.InterestRateItem'
        type: array
    type: object
  handlers.LoanDetailResponse:
    properties:
      amount:
        example: 100000
        type: integer
      approved_by:
        $ref: '#/definitions/handlers.ManagerInfo'
      borrower:
        $ref: '#/definitions/handlers.BorrowerInfo'
      created_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      duration:
        example: 12
        type: integer
      id:
        example: 1
        type: integer
      interest_rate:
        example: 12.5
        type: number
      monthly_payment:
        example: 9000
        type: integer
      outstanding_balance:
        example: 95000
        type: integer
      principal:
        example: 100000
        type: integer
      reason:
        example: Home renovation
        type: string
      status:
        example: Approved
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      phone_number:
        example: "+1234567890"
        type: string
    required:
    - password
    - phone_number
    type: object
  handlers.LoginResponse:
    properties:
      ok:
        example: true
        type: boolean
    type: object
  handlers.MakePaymentRequest:
    properties:
      amount:
        example: 9000
        type: integer
      interest_amount:
        example: 1000
        type: integer
      loan_id:
        example: 1
        type: integer
      principal_amount:
        example: 8000
        type: integer
    required:
    - amount
    - interest_amount
    - loan_id
    - principal_amount
    type: object
  handlers.MakePaymentResponse:
    properties:
      balance_after:
        example: 86000
        type: integer
      ok:
        example: true
        type: boolean
      transaction_id:
        example: TXN-1234567890
        type: string
    type: object
...
  handlers.ManagerLoansResponse:
    properties:
      other_loans:
        items:
          $ref: '#/definitions/handlers.LoanSummary'
        type: array
      requested_loans:
        items:
          $ref: '#/definitions/handlers.LoanSummary'
        type: array
    type: object
  handlers.MeResponse:
    properties:
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: John Doe
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      savings_balance:
        example: 5000
        type: integer
      shares_balance:
        example: 1000
        type: integer
    type: object
  handlers.MemberLoansResponse:
    properties:
      loans:
        items:
          $ref: '#/definitions/handlers.LoanSummary'
        type: array
      monthly_payment_total:
        example: 9000
        type: integer
      total_due:
        example: 95000
        type: integer
    type: object
  handlers.OutstandingLoanItem:
    properties:
      amount:
        example: 50000
        type: integer
      amount_outstanding:
        example: 42000
        type: integer
      amount_repaid:
        example: 10000
        type: integer
      blockchain_hash:
        example: 0xabc123...
        type: string
      blockchain_verified:
        example: true
        type: boolean
      borrower_id:
        example: 45
        type: integer
      borrower_name:
        example: John Doe
        type: string
      borrower_phone:
        example: "+1234567890"
        type: string
      disbursed_date:
        example: "2025-11-01T00:00:00Z"
        type: string
      duration_months:
        example: 12
        type: integer
      expected_repayment_date:
        example: "2026-11-01T00:00:00Z"
        type: string
      interest_rate:
        example: 4
        type: number
      loan_id:
        example: 123
        type: integer
      reason:
        example: Home renovation
        type: string
      status:
        example: Approved
        type: string
      total_repayment:
        example: 52000
        type: integer
    type: object
  handlers.OutstandingLoansResponse:
    properties:
      outstanding_loans:
        items:
          $ref: '#/definitions/handlers.OutstandingLoanItem'
        type: array
      total_count:
        example: 1
        type: integer
      total_outstanding_amount:
        example: 42000
        type: integer
    type: object
  handlers.RequestLoanRequest:
    properties:
      amount:
        example: 100000
        type: integer
      duration:
        example: 12
        type: integer
      reason:
        example: Home renovation
        type: string
    required:
    - amount
    - duration
    - reason
    type: object
  handlers.RequestLoanResponse:
    properties:
      loan_id:
        example: 1
        type: integer
      ok:
        example: true
        type: boolean
    type: object
  handlers.SetInterestRateRequest:
    properties:
      duration_months:
        example: 12
        type: integer
      rate:
        example: 12.5
        type: number
    required:
    - duration_months
    - rate
    type: object
  handlers.UpdateLoanStatusRequest:
    properties:
      status:
        example: Approved
        type: string
    required:
    - status
    type: object
  handlers.UpdateLoanStatusResponse:
    properties:
      message:
        example: Loan approved successfully
        type: string
      ok:
        example: true
        type: boolean
    type: object
  handlers.UserAuditReportResponse:
    properties:
      blockchain_verified_transactions:
        example: 25
        type: integer
      current_outstanding:
        example: 50000
        type: integer
      current_savings_balance:
        example: 80000
        type: integer
      current_shares_balance:
        example: 50000
        type: integer
      joined_date:
        example: "2024-01-15T00:00:00Z"
        type: string
      last_transaction_date:
        example: "2025-12-05T10:00:00Z"
        type: string
      name:
        example: John Doe
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      role:
        example: member
        type: string
      total_deposits:
        example: 100000
        type: integer
      total_loans_amount:
        example: 150000
        type: integer
      total_loans_repaid:
        example: 100000
        type: integer
      total_loans_taken:
        example: 3
        type: integer
      transaction_count:
        example: 25
        type: integer
      user_id:
        example: 45
        type: integer
      verification_rate:
        example: 100
        type: number
    type: object
  handlers.UserDetailResponse:
    properties:
      name:
        example: John Doe
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      savings_balance:
        example: 50000
        type: integer
      shares_balance:
        example: 25000
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
...
  handlers.UserListResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/handlers.UserListItem'
        type: array
    type: object
  handlers.VerifyRequest:
    properties:
      otp:
        example: "123456"
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      request_id:
        example: 123
        type: integer
    required:
    - otp
    - phone_number
    - request_id
    type: object
  handlers.VerifyResponse:
    properties:
      ok:
        example: true
        type: boolean
      session:
        example: abc123def456...
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@8mh.com
    name: API Support
  description: API for 8MH cooperative banking system with blockchain verification
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 8MH API
  version: "1.0"
paths:
...
  /api/v1/audit/blockchain/status:
    get:
      description: Returns overall blockchain integrity and verification statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BlockchainStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get Blockchain Verification Status
      tags:
      - audit
  /api/v1/audit/loans/outstanding:
    get:
      description: Returns all currently outstanding loans with borrower details
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Sort field (amount, created_at, borrower_name)
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.OutstandingLoansResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get Outstanding Loans
      tags:
      - audit
  /api/v1/audit/summary:
    get:
      description: Returns high-level financial totals for audit purposes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.FinancialSummaryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get Financial Summary
      tags:
      - audit
  /api/v1/audit/transactions:
    get:
      description: Returns all transactions for audit trail with filtering options
      parameters:
      - description: Filter by transaction type
        in: query
        name: type
        type: string
      - description: Filter from date (ISO 8601)
        in: query
        name: start_date
        type: string
      - description: Filter to date (ISO 8601)
        in: query
        name: end_date
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Show only blockchain-verified
        in: query
        name: verified_only
        type: boolean
      - description: Records per page (default 100)
        in: query
        name: limit
        type: integer
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuditTransactionsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get All Transactions
      tags:
      - audit
  /api/v1/audit/transactions/export:
    get:
      description: Exports all transactions to Excel or CSV format with blockchain
        verification status
      parameters:
      - description: Filter by transaction type
        in: query
        name: type
        type: string
      - description: Filter from date (ISO 8601)
        in: query
        name: start_date
        type: string
      - description: Filter to date (ISO 8601)
        in: query
        name: end_date
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Export format (excel, csv)
        in: query
        name: format
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      - text/csv
      responses:
        "200":
          description: Excel or CSV file
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Export Transactions to Excel/CSV
      tags:
      - audit
  /api/v1/audit/users/{id}:
    get:
      description: Returns detailed audit report for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserAuditReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get User Audit Report
      tags:
      - audit
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with phone number and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Login with phone and password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Logout and invalidate session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Logout
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Get authenticated user's profile and account information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get current user information
      tags:
      - auth
  /api/v1/auth/verify:
    post:
      consumes:
      - application/json
      description: Verify the OTP code and create an authenticated session
      parameters:
      - description: Verify Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.VerifyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Invalid OTP or expired
          schema:
            $ref: '#/definitions/handlers.VerifyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Verify OTP and create session
      tags:
      - auth
  /api/v1/deposit:
    post:
      consumes:
      - application/json
      description: Manager adds a deposit, creates transaction & block, updates user
        savings balance
      parameters:
      - description: Add Deposit Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddDepositRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AddDepositResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Add a deposit (manager)
      tags:
      - deposits
  /api/v1/home:
    get:
      description: Returns financial statistics based on user role (member, manager,
        auditor)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HomeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get home dashboard statistics
      tags:
      - home
  /api/v1/interest_rates:
    get:
      description: Returns list of all interest rates by duration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.InterestRateResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get all interest rates
      tags:
      - interest-rates
  /api/v1/interest_rates/set:
    post:
      consumes:
      - application/json
      description: Manager sets or updates interest rate for a specific loan duration
      parameters:
      - description: Set Interest Rate Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetInterestRateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Set interest rate for duration (manager)
      tags:
      - interest-rates
  /api/v1/loans/{id}:
    get:
      description: Returns detailed information about a specific loan (manager only)
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoanDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get loan details by ID
      tags:
      - loans
  /api/v1/loans/{id}/update_status:
    post:
      consumes:
      - application/json
      description: Manager approves or rejects a loan. On approval, sets interest
        rate and monthly payment
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateLoanStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpdateLoanStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Update loan status (approve/reject)
      tags:
      - loans
  /api/v1/loans/add:
    post:
      consumes:
      - application/json
      description: Manager directly adds a loan with all details
      parameters:
      - description: Add Loan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddLoanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RequestLoanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Add a new loan directly (manager)
      tags:
      - loans
  /api/v1/loans/manager:
    get:
      description: Returns requested loans first, then all other loans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ManagerLoansResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get all loans (manager view)
      tags:
      - loans
  /api/v1/loans/member:
    get:
      description: Returns list of loans for the authenticated member with summary
        statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MemberLoansResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get member's loans
      tags:
      - loans
  /api/v1/loans/payment:
    post:
      consumes:
      - application/json
      description: Member makes a payment towards their loan, creates transaction
        and blockchain block
      parameters:
      - description: Make Payment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.MakePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MakePaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Make a loan payment
      tags:
      - loans
  /api/v1/loans/request:
    post:
      consumes:
      - application/json
      description: Member requests a loan with amount, duration, and reason
      parameters:
      - description: Loan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RequestLoanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RequestLoanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Request a new loan (member)
      tags:
      - loans
  /api/v1/users:
    get:
      description: Get list of all users, optionally filter by name or phone number
      parameters:
      - description: Search by name or phone number
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: List/Search Users
      tags:
      - users
  /api/v1/users/{id}:
    get:
      description: Get detailed information about a specific user including balances
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get User Details
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"